poc/ray-a2a/requirements.txt
ray[default]>=2.10
pydantic-ai-slim[a2a]>=1.84.1
uvicorn[standard]>=0.34
httpx>=0.28

